When player touches coin: add to player and delete coin.
I was able to get a basic understanding of how the signals work and can explain what I did, if needed.

used the built-in "On Body Enter signal" to detect when a player enters the coin's collision. -- Then using the signal increases the coins value on within the character2d script and print the total coins to the HUD label. -- then the coin is deleted